file split corrected
availability of Send/Connect button corrected for all scenarios,
such as in file tab, when a large file chosen for UDP and then
switching to TCP, remote port entrance, etc.
When a SvrTcp fails to listen, clean up work added, not to block
next round.
log added when routines entered and exited, when verbose = 2 or 3,
except for flow.